app/requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
pytest
flake8
PyGithub
requests
